Create MERN login project
A basic project to create a MERN-based CRUD application, which will validate a user login attempt against a live database via MongoDB Atlas. Additionally, the project will allow new users to create a new set of login credentials for the system, with which they will be able to successfully 'log in'.