Finished login project
Added full back-end support for the following features:
Logging in (including validation) and creating a new account (including validation and checking for existing accounts).

Also finished off the UI, along with a 'Page not found' page for any lost travelers.

Thinking of adding a forgot password feature, but for the time being this project works as intended and should be mobile friendly (If I launched and hosted it somewhere at least).